# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI MySQL Application",
    "description": "台股相關 **FastAPI + MySQL** 後端。\n\n- **互動文件**：本頁 Swagger UI（`/docs`）或 ReDoc（`/redoc`）。\n- **健康檢查**：`GET /health`。\n- **AI 分析**：見標籤「AI 分析」；分階架構可並行呼叫 `raw` 三筆 + `quick-insights` + `final`。\n\n實際部署網域與 CORS 請依環境調整。",
    "version": "1.0.0"
  },
  "paths": {
...
    "/analyze/stock-behavior/basic": {
      "post": {
        "tags": [
          "AI 分析"
        ],
        "summary": "取得股票基本圖表與近期證據資料",
        "operationId": "get_stock_behavior_basic_analyze_stock_behavior_basic_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockBehaviorAnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockBehaviorBasicResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyze/stock-behavior/rag": {
      "post": {
        "tags": [
          "AI 分析"
        ],
        "summary": "取得股票相關 RAG 新聞回覆",
        "description": "Request 只接受 `symbols` 陣列，目前後端只會使用第一個有效股票代號。Response 直接回傳 `news_sources`、`fallback_mode`、`raw_answer`。",
        "operationId": "get_stock_behavior_rag_analyze_stock_behavior_rag_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockBehaviorRagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockBehaviorRagResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyze/stock-behavior/ai": {
      "post": {
        "tags": [
          "AI 分析"
        ],
        "summary": "產生股票 AI 建議分析",
        "description": "Request 只接受 `symbol`、`news_sources`、`fallback_mode`、`raw_answer`。通常可直接把 `/analyze/stock-behavior/rag` 的 response 欄位帶入。",
        "operationId": "get_stock_behavior_ai_analyze_stock_behavior_ai_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockBehaviorAiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockBehaviorAiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalyzeNewsSourceItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "新聞唯一識別碼。",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "新聞標題。",
            "default": ""
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "新聞摘要或內容節錄。",
            "default": ""
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "新聞時間，ISO 8601 格式。",
            "default": ""
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url",
            "description": "原始新聞網址。"
          }
        },
        "type": "object",
        "title": "AnalyzeNewsSourceItem"
      },
...
      "EvidenceUsed": {
        "properties": {
          "price_volume": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Price Volume"
          },
          "chip": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Chip"
          },
          "technical": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Technical"
          },
          "news": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "News"
          },
          "general": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "General"
          }
        },
        "type": "object",
        "title": "EvidenceUsed"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectionPoint": {
        "properties": {
          "day": {
            "type": "integer",
            "title": "Day"
          },
          "relative_price": {
            "type": "number",
            "title": "Relative Price",
            "default": 1.0
          },
          "predicted_close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicted Close"
          },
          "predicted_volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicted Volume"
          },
          "direction": {
            "type": "string",
            "title": "Direction",
            "default": "uncertain"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "day"
        ],
        "title": "ProjectionPoint"
      },
      "RagReferenceAnalysis": {
        "properties": {
          "raw_answer_used_as": {
            "type": "string",
            "enum": [
              "reference_only"
            ],
            "const": "reference_only",
            "title": "Raw Answer Used As",
            "default": "reference_only"
          },
          "rag_sentiment": {
            "type": "string",
            "enum": [
              "bullish",
              "neutral",
              "bearish",
              "mixed",
              "unknown"
            ],
            "title": "Rag Sentiment",
            "default": "unknown"
          },
          "rag_summary": {
            "type": "string",
            "title": "Rag Summary",
            "default": ""
          },
          "news_sources_count": {
            "type": "integer",
            "title": "News Sources Count",
            "default": 0
          },
          "is_confirmed_by_price_volume": {
            "type": "boolean",
            "title": "Is Confirmed By Price Volume",
            "default": false
          },
          "is_confirmed_by_chip": {
            "type": "boolean",
            "title": "Is Confirmed By Chip",
            "default": false
          },
          "is_confirmed_by_technical": {
            "type": "boolean",
            "title": "Is Confirmed By Technical",
            "default": false
          },
          "conflicts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Conflicts"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "RagReferenceAnalysis"
      },
...
      "RiskItem": {
        "properties": {
          "risk_type": {
            "type": "string",
            "title": "Risk Type",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "watch_condition": {
            "type": "string",
            "title": "Watch Condition",
            "default": ""
          }
        },
        "type": "object",
        "title": "RiskItem"
      },
      "ScenarioProjection": {
        "properties": {
          "horizon_days": {
            "type": "integer",
            "title": "Horizon Days",
            "default": 40
          },
          "scenario_key": {
            "type": "string",
            "title": "Scenario Key",
            "default": "primary"
          },
          "scenario_name": {
            "type": "string",
            "title": "Scenario Name",
            "default": "主情境"
          },
          "user_interpretation": {
            "type": "string",
            "title": "User Interpretation",
            "default": ""
          },
          "summary_for_user": {
            "type": "string",
            "title": "Summary For User",
            "default": ""
          },
          "trigger_conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Trigger Conditions"
          },
          "invalidation_conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Invalidation Conditions"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ProjectionPoint"
            },
            "type": "array",
            "title": "Points"
          },
          "line_disclaimer": {
            "type": "string",
            "title": "Line Disclaimer",
            "default": "此趨勢線為 AI 情境推演，非統計預測，不構成投資建議。"
          }
        },
        "type": "object",
        "title": "ScenarioProjection"
      },
...
      "StockBehaviorAiRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "maxLength": 10,
            "minLength": 1,
            "title": "Symbol",
            "description": "單一股票代號，例如 2330。"
          },
          "news_sources": {
            "items": {
              "$ref": "#/components/schemas/AnalyzeNewsSourceItem"
            },
            "type": "array",
            "title": "News Sources",
            "description": "RAG 新聞來源列表，通常直接使用 `/analyze/stock-behavior/rag` 的 response.news_sources。"
          },
          "fallback_mode": {
            "type": "boolean",
            "title": "Fallback Mode",
            "description": "是否為 RAG fallback 模式。通常直接使用 `/analyze/stock-behavior/rag` 的 response.fallback_mode。",
            "default": false
          },
          "raw_answer": {
            "type": "string",
            "title": "Raw Answer",
            "description": "RAG 原始摘要文字。通常直接使用 `/analyze/stock-behavior/rag` 的 response.raw_answer。",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "symbol"
        ],
        "title": "StockBehaviorAiRequest"
      },
      "StockBehaviorAiResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "as_of_date": {
            "type": "string",
            "title": "As Of Date"
          },
          "llm_analysis": {
            "$ref": "#/components/schemas/StockBehaviorAnalysisPayload"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "as_of_date",
          "llm_analysis"
        ],
        "title": "StockBehaviorAiResponse"
      },
      "StockBehaviorAnalysisPayload": {
        "properties": {
          "data_gap": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Data Gap"
          },
          "observations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Observations"
          },
          "inferences": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Inferences"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "current_trend_assessment": {
            "$ref": "#/components/schemas/TrendAssessment"
          },
          "subjective_view": {
            "$ref": "#/components/schemas/SubjectiveView"
          },
          "projection": {
            "$ref": "#/components/schemas/ScenarioProjection"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Risk Level",
            "default": "medium"
          },
          "risk_analysis": {
            "items": {
              "$ref": "#/components/schemas/RiskItem"
            },
            "type": "array",
            "title": "Risk Analysis"
          },
          "rag_reference_analysis": {
            "$ref": "#/components/schemas/RagReferenceAnalysis"
          },
          "evidence_used": {
            "$ref": "#/components/schemas/EvidenceUsed"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "type": "object",
        "title": "StockBehaviorAnalysisPayload"
      },
      "StockBehaviorAnalyzeRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "maxLength": 10,
            "minLength": 1,
            "title": "Symbol"
          },
          "as_of_date": {
            "type": "string",
            "title": "As Of Date",
            "description": "YYYY-MM-DD"
          },
          "horizon_days": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 5.0,
            "title": "Horizon Days",
            "default": 40
          },
          "recent_lookback_days": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 20.0,
            "title": "Recent Lookback Days",
            "default": 365
          },
          "news_lookback_days": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 7.0,
            "title": "News Lookback Days",
            "default": 60
          },
          "max_news_events": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max News Events",
            "default": 10
          },
          "analysis_language": {
            "type": "string",
            "title": "Analysis Language",
            "default": "zh-TW"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "as_of_date"
        ],
        "title": "StockBehaviorAnalyzeRequest"
      },
      "StockBehaviorBasicResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "as_of_date": {
            "type": "string",
            "title": "As Of Date"
          },
          "stored_behavior_profile": {
            "type": "object",
            "title": "Stored Behavior Profile"
          },
          "recent_evidence": {
            "type": "object",
            "title": "Recent Evidence"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "as_of_date"
        ],
        "title": "StockBehaviorBasicResponse"
      },
      "StockBehaviorRagRequest": {
        "properties": {
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Symbols",
            "description": "股票代號陣列。Swagger 與目前後端流程只會使用第一個有效代號。",
            "examples": [
              [
                "2330"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "symbols"
        ],
        "title": "StockBehaviorRagRequest"
      },
      "StockBehaviorRagResponse": {
        "properties": {
          "news_sources": {
            "items": {
              "$ref": "#/components/schemas/AnalyzeNewsSourceItem"
            },
            "type": "array",
            "title": "News Sources",
            "description": "RAG 回傳的新聞來源列表。"
          },
          "fallback_mode": {
            "type": "boolean",
            "title": "Fallback Mode",
            "description": "是否啟用 fallback 模式。",
            "default": false
          },
          "raw_answer": {
            "type": "string",
            "title": "Raw Answer",
            "description": "RAG 回傳的原始摘要文字。",
            "default": ""
          }
        },
        "type": "object",
        "title": "StockBehaviorRagResponse"
      },
      "SubjectiveView": {
        "properties": {
          "opinion": {
            "type": "string",
            "title": "Opinion",
            "default": ""
          },
          "supported_evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Evidence"
          },
          "invalidation_conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Invalidation Conditions"
          }
        },
        "type": "object",
        "title": "SubjectiveView"
      },
...
      "TrendAssessment": {
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "bullish",
              "mildly_bullish",
              "neutral",
              "mildly_bearish",
              "bearish",
              "uncertain"
            ],
            "title": "State",
            "default": "uncertain"
          },
          "confidence_level": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Confidence Level",
            "default": "low"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          }
        },
        "type": "object",
        "title": "TrendAssessment"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}